@@ -5,6 +5,7 @@
   "main": "lib/index.js",
   "module": "lib/index.esm.js",
   "umd:main": "lib/index.umd.js",
+  "types": "lib/index.d.ts",
   "source": "src/index.ts",
   "repository": {
     "type": "git",
@@ -20,21 +21,20 @@
   "scripts": {
     "test": "jest",
     "test:ci": "jest -u --coverage",
+    "format": "prettier --write \"src/**/*.[jt]s\"",
+    "lint": "eslint \"src/**/*.[jt]s\"",
     "build:js": "microbundle",
     "build": "npm run build:js && npm run test",
     "prepare": "npm run build",
     "release": "semantic-release --debug"
   },
-  "husky": {
-    "hooks": {
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
-    }
-  },
   "jest": {
     "transform": {
       "^.+\\.tsx?$": "ts-jest"
     },
-    "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
+    "testMatch": [
+      "**/__tests__/**/*-test.[jt]s?(x)"
+    ],
     "moduleFileExtensions": [
       "ts",
       "tsx",
@@ -52,33 +52,33 @@
     "testEnvironment": "node"
   },
   "devDependencies": {
-    "@commitlint/cli": "11.0.0",
-    "@commitlint/config-conventional": "11.0.0",
+    "@ethersproject/contracts": "5.1.0",
     "@semantic-release/changelog": "5.0.1",
     "@semantic-release/git": "9.0.0",
-    "@types/jest": "26.0.20",
+    "@types/elliptic": "6.4.12",
+    "@types/jest": "26.0.22",
+    "@typescript-eslint/eslint-plugin": "4.22.0",
+    "@typescript-eslint/parser": "4.22.0",
     "codecov": "3.8.1",
+    "eslint": "7.24.0",
+    "eslint-config-prettier": "8.2.0",
     "ganache-cli": "6.12.2",
     "jest": "26.6.3",
     "microbundle": "0.13.0",
     "prettier": "2.2.1",
-    "semantic-release": "17.3.9",
-    "truffle-contract": "4.0.31",
-    "ts-jest": "26.5.1",
-    "ts-lint": "4.5.1",
-    "tslint-config-prettier": "1.18.0",
-    "tslint-eslint-rules": "5.4.0",
-    "typescript": "4.1.5",
-    "web3": "1.3.4"
+    "semantic-release": "17.4.2",
+    "ts-jest": "26.5.4",
+    "typescript": "4.2.4"
   },
   "dependencies": {
+    "@ethersproject/abstract-signer": "^5.1.0",
+    "@ethersproject/address": "^5.1.0",
+    "@ethersproject/transactions": "^5.1.0",
+    "@ethersproject/wallet": "^5.1.0",
     "buffer": "^6.0.3",
-    "did-jwt": "^4.8.0",
-    "did-resolver": "^2.1.2",
-    "ethjs-contract": "^0.2.3",
-    "ethjs-provider-http": "^0.1.6",
-    "ethjs-query": "^0.3.8",
+    "did-jwt": "^5.1.2",
+    "did-resolver": "^3.1.0",
     "ethr-did-registry": "0.0.3",
-    "ethr-did-resolver": "^3.0.3"
+    "ethr-did-resolver": "^4.1.0"
   }
 }
